Refactor scrambling logic and prepare for giving a warning when attacking a territory that can be scrambled.
